Use a binary search to fill TextChunks
`TextChunks::next_chunk` is O(n) over the larger of the desired chunk
size or the input size. This is especially apparent when using a
relatively slow tokenizer like tiktoken-rs and requesting a large
chunk size.

Here we change `next_chunk` to use a binary search to fill the chunk
from semantic sections, such that the largest possible chunk satisfying
the capacity requirements is returned on each iteration. This
implementation is O(log n) over chunk size.

This slightly changes the semantics of the chunker where a chunk size
range is supplied by the caller, the original implementation would
return the smallest possible chunk that satisfies the requirements, this
will return the largest possible chunk.