feat!: Update Splitters to require a ChunkConfig, and trim by default
The configuration settings for chunks were split between construction (sizer and trim) and chunk creation (capacity).

To make way for settings like chunk overlap, it seemed best to pull all of this into one place, because that will introduce invariants (chunk overlap cannot be bigger than chunk size). By creating ChunkConfig, it allows for one place to put all of the configuration to describe the chunking method, and allow for validating that is is a valid chunk config.

The downside is that a splitter can only ever have one config, but in practice almost every use I saw was doing this anyway. For each use case, you want to define all of these values together to describe the chunks you want.

Also, this changes the default behavior to trim by default. Every use case I saw in practice turned this setting on, and it makes sense to have this by default.