feat!: Use shared fallback logic across all splitters
This changes unicode segmentation to be used as a fallback for all splitter types.

Ideally we are utilizing the custom logic anyway. We only use these defaults to make sure the cursor is still able to move forward.

This does improve runtime performance because we can skip several smaller levels by default, only using them if necessary.

This also causes some change to markdown behavior (hence the breaking change) because prior to this it was prioritizing sentence segmentation over some of the smaller levels in Markdown, which caused some strange break points. So this behavior should be more correct.