Optimization for SemanticSplitRange searching
Leverages the fact that these ranges are guaranteed to be sorted now. Rather than doing a retain, which likely moves items in the vec and also has to iterate over all values, instead there is a cursor to extract the desired slice, and we can move the cursor to the first item that would still be in the range we desire.